+++ src/stat.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: stat.cc,v 1.350 2001/10/24 06:55:44 hno Exp $
+ * $Id: stat.cc,v 1.351 2001/10/24 08:19:08 hno Exp $
  *
  * DEBUG: section 18    Cache Manager Statistics
  * AUTHOR: Harvest Derived
@@ -1426,7 +1426,7 @@ statClientRequests(StoreEntry * s)
 	storeAppendPrintf(s, "log_type %s\n", log_tags[http->log_type]);
 	storeAppendPrintf(s, "out.offset %ld, out.size %lu\n",
 	    (long int) http->out.offset, (unsigned long int) http->out.size);
-	storeAppendPrintf(s, "req_sz %d\n", http->req_sz);
+	storeAppendPrintf(s, "req_sz %ld\n", (long int) http->req_sz);
 	e = http->entry;
 	storeAppendPrintf(s, "entry %p/%s\n", e, e ? storeKeyText(e->hash.key) : "N/A");
 	e = http->old_entry;
