+++ src/store_dir.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store_dir.cc,v 1.134 2001/08/16 00:16:18 hno Exp $
+ * $Id: store_dir.cc,v 1.135 2001/10/24 08:19:09 hno Exp $
  *
  * DEBUG: section 47    Store Directory Routines
  * AUTHOR: Duane Wessels
@@ -290,8 +290,8 @@ storeDirStats(StoreEntry * sentry)
     storeAppendPrintf(sentry, "Store Directory Statistics:\n");
     storeAppendPrintf(sentry, "Store Entries          : %d\n",
 	memInUse(MEM_STOREENTRY));
-    storeAppendPrintf(sentry, "Maximum Swap Size      : %8d KB\n",
-	Config.Swap.maxSize);
+    storeAppendPrintf(sentry, "Maximum Swap Size      : %8ld KB\n",
+	(long int) Config.Swap.maxSize);
     storeAppendPrintf(sentry, "Current Store Swap Size: %8d KB\n",
 	store_swap_size);
     storeAppendPrintf(sentry, "Current Capacity       : %d%% used, %d%% free\n",
